+++ CHANGELOG.md
@@ -3,6 +3,14 @@
 
 Full changelog for PHP Quill Renderer
 
+## v0.80.0 
+
+* I'm now only testing against PHP 7+. [Tests]
+* Added tests for setting attributes. [Tests]
+* Switched to PSR4
+* Minor change to API if using Quill (Render) class, after the PSR4 change I didn't like Quill/Quill.
+* Basic support for images (outputting the base64 directly via <img> src="")
+
 ## v0.70.0 - 2017-04-19
 
 * Added the ability to set the HTML tag for the following Quill attributes, bold, italic, script, strike and underline. [Feature]